Implement TCPServer#listen.
The JDK socket classes do both bind and listen at once and do not
expose a way to change the backlog after that. To work around
this, I reflectively attempt to access the class in which they
bind the native listen(2) function. If I'm able to use it, we will
set the backlog. If I'm not, TCPServer#listen remains mostly a
no-op.

Note that a strategy like RubySocket, where we defer the creation
of the actual socket, might work here...but that has not yet been
proven out, and subsequent calls to listen would not work.